requirements.txt
python-telegram-bot[ext]==21.10
python-dotenv==1.1.0
aiosqlite==0.20.0
